Use Stdlib instead of Pervasives
Pervasives was deprecated in favor of directly using Stdlib in 4.08 and
then removed in the upcoming 5.0. This change makes javalib buildable
with the multi-core ocaml 5.0~alpha.

NOTE: I checked that the project builds with 5.0 but I didn't try to run
anything. In particular I'm not sure what jvmti native stubs do.